third_party/WebKit/LayoutTests/fast/html/imports/import-defaultView.html

<!DOCTYPE html>
<html>
<script src="../../../resources/js-test.js"></script>
<head>
<link id="importLink" rel="import" href="resources/hello.html">
</head>
<body>
<script>
// This used to be non-null but not longer is.
// See http://crbug.com/408272
shouldBeNull("importLink.import.defaultView");

window.jsTestIsAsync = true;

var xhr = new XMLHttpRequest();
xhr.open("GET", "resources/root.html", true);
xhr.responseType = "document";
xhr.onload = function() {
  shouldBeNull("xhr.response.defaultView");

  createdDoc = document.implementation.createHTMLDocument("Hello");
  shouldBeNull("createdDoc.defaultView");
  finishJSTest();
};
xhr.send();

</script>
</body>
</html>
